backend/requirements.txt
flask
flask-cors
ultralytics
Pillow
numpy
gunicorn
opencv-python
